==> 02_Scripts/STEP_01_process_genotypes.R <==

# ------------------------------------ #
# ------------- STEP 01  ------------- #
# ------------------------------------ #

### process SNP genotypes
### Author: Annabel Smith & Di Binyin

# Load workspace:
load("buffel_burning.RData")

# load functions:
invisible(lapply(paste("01_Functions/",dir("01_Functions"),sep=""),function(x) source(x)))

#  IMPORT & FORMAT DATA:    	# ----

# dir with original SNP data:
data.dir<-"/Users/annabelsmith/Documents/01_Current/PROJECTS/02_FIRESCAPE/DATA/01_FIRESCAPE_sample_data/DHype_DCen_original_data_DO_NOT_MODIFY/Report-DCen19-4425"
dir(data.dir)

### ~~~ SINGLE ROW SNP data ~~~ ###

# The first six rows contain metadata:
gt_onerow<-read.csv(paste(data.dir,"Report_DCen19-4425_SNP_singlerow_2_.csv",sep="/"),header=T,skip=6)
ghead(gt_onerow); dim(gt_onerow)

### ---- LOCUS INFO ---- ###

# WARNING: numeric subset - check with raw data
# First 17 columns have locus info (DPlan):
# First 34 columns have locus info (DCen19):
loc_info_all<-colnames(gt_onerow[,1:34])

linf<-gt_onerow[,loc_info_all]
linf<-tidy.df(linf)
linf<-data.frame(locus=paste("L",1:nrow(linf),sep=""),linf)
linf$locus<-as.character(linf$locus)
head(linf); dim(linf)

head(linf[,1:5])
head(linf[,20:30])
head(linf[,30:35])
hist(linf$CallRate)

check.rows(linf)

### ---- END LOCUS INFO ---- ###

# Indices for SNPs cols, containing individual genotypes:
snp_cols<-colnames(gt_onerow)[-which(colnames(gt_onerow) %in% loc_info_all)]
head(snp_cols)
length(snp_cols)

# Make locus columns characters:
gt_onerow[,snp_cols]<-apply(gt_onerow[,snp_cols],2,as.character)

# Replace - with NA:
gt_onerow[,snp_cols]<-as.data.frame(apply(gt_onerow[,snp_cols],2,function(x)gsub("-","NA",x)))

# Make locus columns numeric:
# (gives warning: NAs introduced by coercion)
gt_onerow[,snp_cols]<-apply(gt_onerow[,snp_cols],2,as.numeric)
ghead(gt_onerow)

# SNP data with individuals on rows:
# WARNING: character subset - check with raw data
# DPlan, DCen19 = "RepAvg":
snp_onerow<-data.frame(t(gt_onerow[,(which(colnames(gt_onerow)=="RepAvg")+1):length(gt_onerow)]))
colnames(snp_onerow)<-linf$locus
ind.names<-rownames(snp_onerow)

# Add individuals names back in:
snp_onerow<-data.frame(ind=ind.names,snp_onerow)
ghead(snp_onerow); dim(snp_onerow)

# Add site:
snp_onerow$site<-substr(snp_onerow$ind,1,unlist(gregexpr("_", snp_onerow$ind))-1)

# Tidy:
snp_onerow<-tidy.df(snp_onerow)

# Reorder:
snp_onerow<-snp_onerow[,c(which(colnames(snp_onerow) %in% c("site","ind"))[c(2,1)],grep("L",colnames(snp_onerow)))]
snp_onerow<-snp_onerow[order(snp_onerow$site,snp_onerow$ind),]
snp_onerow<-tidy.df(snp_onerow)
snp_onerow$site<-as.factor(snp_onerow$site)
ghead(snp_onerow); dim(snp_onerow)

# For Cenchrus X01u_03.1 is the replicate sample

# save.image("binyin_winter.RData")

# onerow == alleles coded using dartseq notation:
# 0 = Reference allele homozygote AA
# 1 = SNP allele homozygote GG
# 2 = heterozygote AG

### ~~~ END FIX SITE COLS ~~~ ###

# Import site data:
sdat<-read.table(paste("00_data/Cenchrus_site_data.txt",sep=""),header=T)
sdat<-sdat[sdat$sequenced==1,]
sdat<-tidy.df(sdat)
head(sdat); dim(sdat)

# close import ----

ghead(snp_onerow); dim(snp_onerow)
head(sdat); dim(sdat)

# number of site in genetic data should == number of sites in site data:
length(levels(snp_onerow$site))==dim(sdat)[1]

# The result from IMPORT & FORMAT DATA includes ALL 69799 SNP loci; all individuals that were sequenced, including the replicate sample: X01u_03.1

#  Check & remove duplicate genotypes   ----

ghead(snp_onerow); dim(snp_onerow)
snp_onerow[,1:2]

# Remove dups and save file with all duplicate samples for checking. All dups indicated by ".":
dup_samples<-unique(substr(snp_onerow$ind[grep("\\.",snp_onerow$ind)],1,unlist(gregexpr("\\.",snp_onerow$ind[grep("\\.",snp_onerow$ind)]))-1))
ind_all<-snp_onerow$ind
ind_all[grep("\\.",snp_onerow$ind)]<-substr(snp_onerow$ind[grep("\\.",snp_onerow$ind)],1,unlist(gregexpr("\\.",snp_onerow$ind[grep("\\.",snp_onerow$ind)]))-1)
# check
ch.dat1<-data.frame(ind_all,snp_onerow$ind)
head(ch.dat1)
ch.dat1[which(!ch.dat1$ind_all==ch.dat1$snp_onerow.ind),]

# Create df for all dups:
dup_data<-snp_onerow[ind_all %in% dup_samples,]
dup_data<-tidy.df(dup_data)
ghead(dup_data)

# Remove dups from main data:
snp_onerow<-snp_onerow[-grep("\\.",snp_onerow$ind),]
snp_onerow<-tidy.df(snp_onerow)
ghead(snp_onerow); dim(snp_onerow)

dloci<-colnames(dup_data)[grep("L",colnames(dup_data))]
ghead(dup_data); dim(dup_data)

# What percentage of loci do not match in the duplicate data? 

dup_inds<-unique(levels(dup_data$site))
dup_res<-data.frame(ind=dup_inds,mistyped=NA,correct_typed=NA)

for (i in 1:length(dup_inds)){

ind.thisrun<-dup_inds[i]

data.thisrun<-dup_data[grep(ind.thisrun,dup_data$ind),-as.numeric(unlist(apply(dup_data[grep(ind.thisrun,dup_data$ind),3:length(dup_data)],1,function(x)which(is.na(x))))+2)]
data.thisrun[,c(1,2,sample(3:length(data.thisrun),size=10))]
ghead(data.thisrun)

t.thisrun<-table(apply(data.thisrun[,3:length(data.thisrun)],2,function(x)length(unique(x))))

dup_res[i,2]<-t.thisrun[2]
dup_res[i,3]<-t.thisrun[1]

}

dup_res$prop_mistyped<-dup_res$mistyped/(dup_res$mistyped+dup_res$correct_typed)
dup_res
head(linf,2) 

# close dups ----

#     RESULT = FULL DATA SET:   	 ####

# The full data set includes ALL 69799 loci (plus two columns for site and individual); all 93 sequenced individuals and no dups. 
ghead(snp_onerow); dim(snp_onerow)

# dup_data includes 2 genotypes for one indiviudal
ghead(dup_data); dim(dup_data)

# onerow == alleles coded using dartseq notation:
# 0 = Reference allele homozygote
# 1 = SNP allele homozygote
# 2 = heterozygote

# It is wrong to code presence/absences as alleles - they indicate the opposite of the true genotype. E.g. 0,1 indicates that only the SNP allele is present, i.e. a SNP homozygote; 1,1 indicates both alleles are present, i.e. a heterozygote. If these were coded as alleles 0,1 would look like a heterozygote and 1,1 a homozygote but the opposite is true. As such, this is critical.
ghead(snp_onerow); dim(snp_onerow)

# Site data includes all 19 sites at which individuals were sequenced
head(sdat)

# All locus info:
head(linf)

# close full ####

# save.image("binyin_winter.RData")

#  IDENTIFY DUPLICATE SEQUENCES   #
#   	AND FORMAT LOCUS DATA:  	 #
# ----

# The following "things to consider" was written for Plantago lanceolata in which multiple rounds of sequencing produced duplicate sequences. I don't think this is a problem for C. cenchrus because we're only dealing with one sequencing run. But I'm going to check to make sure. 

# Things to consider:

# There are duplicated sequences; these can be removed, even if they're different SNPs, to reduce linkage
# Even after removing duplicated sequences, there are duplicated loci because of single base pair differences (e.g. the same AlleleID but one bp difference in the sequence because of genotyping error)
# Not all duplicated sequences have the same AlleleID, so duplicated TrimmedSequences AND duplicated AlleleIDs will need to be removed
# However, even after these steps are taken, we still have duplicates. Sometimes a SNP was scored in one direction and again in the other. E.g. it was scored as a G>A SNP on one row and a A>G SNP on another row. These will, by definition, have different sequences, even if there are no genotyping errors. They can also have different AlleleIDs. So use blastn to get a similarity score for each locus. 
# If we simply remove the second record of a duplicated sequence or AlleleID, we will lose information, because there are many cases where the first has a very low call rate and the second has a much higher call rate. So, before removing duplicates, the loci will need to be ordered by call rate, or some other quality metric, to ensure the best ones are kept. 

head(linf[,1:7],2)

# some programs (e.g. bowtie2) read the > character as a separator, so update the allele ID so that the sequences can be identified in the output:
seq2<-linf[,c("AlleleID","TrimmedSequence")]
seq2$AlleleID<-as.character(seq2$AlleleID)
seq2$TrimmedSequence<-as.character(seq2$TrimmedSequence)
seq2$ai2<-seq2$AlleleID

# get first part of the seq ref for the ID:
seq2$ai3<-substr(seq2$ai2,1,unlist(gregexpr("F",seq2$ai2))+1)
seq2$ai3<-substr(seq2$ai3,1,nchar(seq2$ai3)-1)
seq2$ai3<-substr(seq2$ai3,1,nchar(seq2$ai3)-2)
head(seq2)

# get SNP position
seq2$pos<-lapply(seq2$ai2,function(x)substr(x,unlist(gregexpr("-",x))[[1]]+1,unlist(gregexpr(":",x))[[1]]-1))

# get SNP type:
seq2$snp1<-substr(seq2$ai2,(nchar(seq2$ai2)-2),(nchar(seq2$ai2)-2))
seq2$snp2<-substr(seq2$ai2,(nchar(seq2$ai2)),(nchar(seq2$ai2)))
seq2$snp<-paste(seq2$snp1,seq2$snp2,sep="_")

# And make new allele ID:
seq2$ai4<-paste(seq2$ai3,seq2$pos,seq2$snp,sep="_")
head(seq2)
check.rows(seq2)

length(which(duplicated(seq2$TrimmedSequence)))

# This should be zero:
# If it's not, there's some code in the DPlan script for fixing this. 
length(which(duplicated(seq2$ai4)))
head(seq2)

# Write fasta file with all sequences:

for (i in 1:nrow(seq2)){

data.thisrun<-seq2[i,]

write(paste(">",data.thisrun$ai4,sep=""),file="DCen19_seqs.fa",append=T)
write(data.thisrun$TrimmedSequence,file="DCen19_seqs.fa",append=T)

} # close for

# Import blast results:
blast_dir<-"RESULTS/BLAST"
dir(blast_dir)

# Select lines that bound data:
start_line<-grep("hits",readLines(paste(blast_dir,"DCen19_result.out",sep="/")))+1
end_line<-grep("BLASTN 2.7.1+",readLines(paste(blast_dir,"DCen19_result.out",sep="/")))-1
end_line<-c(end_line[2:length(end_line)],(length(readLines(paste(blast_dir,"DCen19_result.out",sep="/")))-1))
lines_to_read<-unlist(mapply(seq,from=start_line,to=end_line))

# Read data lines:
blast_res<-strsplit(readLines(paste(blast_dir,"DCen19_result.out",sep="/"))[lines_to_read],split="\t")
blast_res<-data.frame(do.call(rbind,blast_res))
colnames(blast_res)<-c("query_acc_ver", "subject_acc_ver", "perc_identity", "alignment_length", "mismatches", "gap_opens", "q_start", "q_end", "s_start", "s_end", "evalue"," bit_score")
# gives NA warning
blast_res<-data.frame(blast_res[,1:2],apply(blast_res[,3:ncol(blast_res)],2,function(x) as.numeric(as.character(x))))

# Remove rows where the match is with itself:
blast_res<-blast_res[,1:3]
blast_res$same<-ifelse(as.character(blast_res$query_acc_ver)==as.character(blast_res$subject_acc_ver),1,0)
blast_res<-blast_res[which(blast_res$same==0),]

# linf and seq2 have not been ordered, so if their AlleleIDs still match, the new ai4 can be added directly to linf:
dim(linf)[1]==dim(seq2)[1]
table(linf$AlleleID==seq2$AlleleID)

linf<-data.frame(locus=linf$locus,ai4=seq2$ai4,linf[,which(colnames(linf)=="AlleleID"):ncol(linf)])
head(linf,2); dim(linf)

length(which(duplicated(linf$AlleleID)))
length(which(duplicated(linf$ai4)))

# Order by call rate, so the best of the duplicates will be kept:
linf2<-linf[order(-linf$CallRate),]
linf2<-tidy.df(linf2)

# Identify direct duplicate sequences:
dup_seq<-as.character(linf2$locus[which(duplicated(linf2$TrimmedSequence))])
head(dup_seq); length(dup_seq)

# Identify duplicate AlleleID:
# Some remain from single bp differences
# This should be zero for DCen
dup_aid<-as.character(linf2$locus[which(duplicated(linf2$AlleleID))])
head(dup_aid)

# Update linf2, removing the straight-forward dups:
strfwd<-unique(c(dup_seq))
linf2<-linf2[-which(linf2$locus %in% strfwd),]
linf2<-tidy.df(linf2)
head(linf2, 3); dim(linf2)

# (3) Remove significant matches from blast:
blast_res$query_acc_ver<-as.character(blast_res$query_acc_ver)
blast_res$subject_acc_ver<-as.character(blast_res$subject_acc_ver)
all.matchloci<-unique(c(as.character(blast_res$query_acc_ver),as.character(blast_res$subject_acc_ver)))

# Many of the blast results have already been removed through dup sequences and AlleleIDs, so can simplify the blast results:
all.matchloci<-all.matchloci[which(all.matchloci %in% linf2$ai4)]
blast_res<-blast_res[unique(c(which(blast_res$query_acc_ver %in% all.matchloci),which(blast_res$subject_acc_ver %in% all.matchloci))),]
blast_res<-tidy.df(blast_res)
head(blast_res); dim(blast_res)

removed.loci<-list()

# Very slow, several minutes

for (i in 1:nrow(linf2)){

ai.thisrun<-as.character(linf2$ai4[i])

# If it's unique, it won't appear in the blast results; skip to next
if(ai.thisrun %in% all.matchloci==F) next

if(length(which(blast_res$query_acc_ver %in% ai.thisrun))>0) lines_i<-which(blast_res$query_acc_ver %in% ai.thisrun)
if(length(which(blast_res$subject_acc_ver %in% ai.thisrun))>0) lines_j<-which(blast_res$subject_acc_ver %in% ai.thisrun)

blast.thisrun<-blast_res[c(lines_i,lines_j),]
matches<-unique(c(blast.thisrun$query_acc_ver, blast.thisrun$subject_acc_ver))

# The data containing matches:
linf.thisrun<-linf2[which(linf2$ai4 %in% matches),]

# Sometimes the match(es) will already have been removed, so skip if this is the case:
if(nrow(linf.thisrun)==1) next

if(nrow(linf.thisrun)>1) print(paste("at i = ",i," there are ",nrow(linf.thisrun)," lines; performing removal",sep=""))

# This is ordered by call rate, so remove all of the loci that appear below the first
l.toremove<-as.character(linf.thisrun$ai4[2:nrow(linf.thisrun)])

# Which of the loci to remove have not already been added?
if(length(which(l.toremove %in% unlist(removed.loci)==F))>0) l.toremove<-l.toremove[which(l.toremove %in% unlist(removed.loci)==F)] else l.toremove<-NULL

if(length(l.toremove)>0) removed.loci[[i]]<-l.toremove

if(length(l.toremove)>0) print (paste("removing the following loci: ", l.toremove, sep=""))

} # close for 

save.image("binyin_winter.RData")

loc_res<-unlist(removed.loci)
head(loc_res,10); length(loc_res)
dup_blast<-as.character(linf2$locus[linf2$ai4 %in% loc_res])

# Add all dups to the main locus info data frame:
head(dup_seq); length(dup_seq)
head(dup_aid)
head(dup_blast); length(dup_blast)

all_dups<-unique(c(dup_seq,dup_blast))
linf$duplicate<-ifelse(linf$locus %in% all_dups,1,0)
head(linf,2); dim(linf)
check.rows(linf)
table(linf$duplicate)

# close sequence and locus format----
